+++ js/until.js
@@ -1,0 +1,3 @@
+const getRandomInt = (min, max) => Math.floor(Math.random) * (max - min + 1) + min;
+
+export {getRandomInt};
